In this commit I add a method to obtain the previous state
fsm_get_previous_state, previous

as well as arguments for enter and leave these are:

enter (this, previous)
where this is the current state and previous is the previous state, in some previous will be undefined.

leave(this, next)
where this is the current state and next is the next state.

I fixed a small problem with the readme where it still kept elements from when the triggers were transitions.

and added a new macro to indicate the current version.